# public/swagger/swagger.yml
openapi: 3.0.0
info:
  title: 'TiSaude API'
  version: '0.1'
paths:
...
  /api/doctors:
    get:
      tags:
        - Médicos
      summary: 'Lista de Médicos'
      description: 'Lista de Médicos'
      operationId: 4479c235c75c92e2425c15211470f478
      responses:
        '200':
          description: 'Lista de Médicos'
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Médicos
      summary: 'Cria um novo Médico'
      description: 'Cria um novo Médico'
      operationId: ee9dd1ffdbcab8d6d47b2dd0bd06ed9b
      parameters:
        -
          name: name
          in: query
          description: Nome
          required: true
      responses:
        '200':
          description: 'Cria um novo Médico'
      security:
        -
          bearerAuth: []
  '/api/doctors/{id}':
    get:
      tags:
        - Médicos
      summary: 'Exibe um Médico'
      description: 'Exibe um Médico'
      operationId: d38f21421c3eca3294015240d2d7e1ad
      parameters:
        -
          name: id
          in: path
          description: 'ID do Médico'
          required: true
      responses:
        '200':
          description: 'Exibe um Médico'
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Médicos
      summary: 'Atualiza um Médico'
      description: 'Atualiza um Médico'
      operationId: dd6e159937989a95449bd7ebab306776
      parameters:
        -
          name: id
          in: path
          description: 'ID do Médico'
          required: true
        -
          name: name
          in: query
          description: Nome
          required: true
      responses:
        '200':
          description: 'Atualiza um Médico'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Médicos
      summary: 'Deleta um Médico'
      description: 'Deleta um Médico'
      operationId: 04bc41e2ab78cd139c19896955c012af
      parameters:
        -
          name: id
          in: path
          description: 'ID do Médico'
          required: true
      responses:
        '200':
          description: 'Deleta um Médico'
      security:
        -
          bearerAuth: []
  '/api/doctors/{id}/add-specialties':
    post:
      tags:
        - Médicos
      summary: 'Adiciona uma ou várias especialidades a um Médico'
      description: 'Adiciona uma ou várias especialidades a um Médico'
      operationId: c6d845b351eadd89376732277c8c73bd
      parameters:
        -
          name: id
          in: path
          description: 'ID do Médico'
          required: true
        -
          name: 'specialty_ids[]'
          in: query
          description: 'IDs das Especialidades'
          required: true
          schema:
            type: array
            items:
              type: integer
            minItems: 1
      responses:
        '200':
          description: 'Especialidades adicionadas'
      security:
        -
          bearerAuth: []
  '/api/doctors/{id}/remove-specialties':
    post:
      tags:
        - Médicos
      summary: 'Remove uma especialidade de um Médico'
      description: 'Remove uma especialidade de um Médico'
      operationId: 545fee247760e172a6fc982d60a25995
      parameters:
        -
          name: id
          in: path
          description: 'ID do Médico'
          required: true
        -
          name: specialty_id
          in: query
          description: 'ID da Especialidade'
          required: true
      responses:
        '200':
          description: 'Especialidade removida'
      security:
        -
          bearerAuth: []
  '/api/doctors/{id}/appointments':
    get:
      tags:
        - Médicos
      summary: 'Exibe as consultas de um Médico'
      description: 'Exibe as consultas de um Médico'
      operationId: 5ee5f439de5bb850a0848a40886ae446
      parameters:
        -
          name: id
          in: path
          description: 'ID do Médico'
          required: true
      responses:
        '200':
          description: 'Consulta do Médico'
      security:
        -
          bearerAuth: []
...
  /api/pacients:
    get:
      tags:
        - Pacientes
      summary: 'Lista de Pacientes'
      description: 'Lista de Pacientes'
      operationId: 3549af4e35db6d775f55ecf368fec281
      responses:
        '200':
          description: 'Lista de Pacientes'
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Pacientes
      summary: 'Cria um novo Paciente'
      description: 'Cria um novo Paciente'
      operationId: 56efce158a5c319d998932b7a6ceec84
      parameters:
        -
          name: name
          in: query
          description: Nome
          required: true
      responses:
        '200':
          description: 'Cria um novo Paciente'
      security:
        -
          bearerAuth: []
  '/api/pacients/{id}':
    get:
      tags:
        - Pacientes
      summary: 'Exibe um Paciente'
      description: 'Exibe um Paciente'
      operationId: 2ea05843b7246dcc33e2df47ae72f2e2
      parameters:
        -
          name: id
          in: path
          description: 'ID do Paciente'
          required: true
      responses:
        '200':
          description: 'Exibe um Paciente'
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Pacientes
      summary: 'Atualiza um Paciente'
      description: 'Atualiza um Paciente'
      operationId: 6631e4ca15795ee7c1bea2ffbf998558
      parameters:
        -
          name: id
          in: path
          description: 'ID do Paciente'
          required: true
        -
          name: name
          in: query
          description: Nome
          required: true
      responses:
        '200':
          description: 'Atualiza um Paciente'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Pacientes
      summary: 'Deleta um Paciente'
      description: 'Deleta um Paciente'
      operationId: 8ce10996f68a2007ba5cf78c2b96cd0d
      parameters:
        -
          name: id
          in: path
          description: 'ID do Paciente'
          required: true
      responses:
        '200':
          description: 'Deleta um Paciente'
      security:
        -
          bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: apiKey
      name: Authorization
      in: header